+++ CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.2.2] - 2024-10-01 
+
+### Changes
+
+- Token select dialog upgrade (Search, Select, Scroll and Add!)
+- Homepage is visible, even when disconnected &/ without a wallet
+- About section added on homepage
+
+### BonePools Activated
+
+- DGOne
+
+### Bugfixes
+
+- Could not find the add button on token ui dialog
+- Xatte'R' branding
+
 ## [2.2.1] - 2024-09-30 
 
 ### Changes